requirements.txt
fastapi[standard]
sqlalchemy
uvicorn
python-dotenv
mysql
pytest
httpx
pydantic
passlib[bcrypt]
python-jose[cryptography]
structlog
Pillow
